requirements.txt
Django==2.2.3
django-taggit==1.1.0
django-taggit-serializer==0.1.7
djangorestframework==3.9.4
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
